macos compile fix
On macos, the template specialization 'std::ctype<unsigned char>' is
undefined, however std::isalnum with unsigned char and a locale requires this.
Also, std::ctype<> is able to classify any char, not just the
non-negative ones as is the case with the C99 function taking an int argument.

So just drop the type cast to unsigned char, and it should work
everywhere.